auth context changes , header refresh when name changes , credit side drawer new changes
